requirements.txt
streamlit
st_lottie
PyPDF2
io
openai
docx2txtpyperclip
